requirements.txt
requests
pandas
yfinance
python-telegram-bot
python-dotenv
